refactor!: move identity structs to a unified module tree
This breaks compatibility because the `wasm_bindgen` macro shadows
the `wasm_bindgen` passive attribute, which means that we can't
conditionally enable the `wasm_bindgen(js_name)` attributes.
