+++ crypto-ffi/src/lib.rs
@@ -40,6 +40,7 @@ mod credential_type;
 mod crl;
 mod decrypted_message;
 mod error;
+mod identity;
 mod metadata;
 
 pub use bundles::{
@@ -58,4 +59,5 @@ pub use error::proteus::ProteusError;
 pub use error::{CoreCryptoResult, core_crypto::CoreCryptoError, mls::MlsError};
 #[cfg(target_family = "wasm")]
 pub use error::{WasmCryptoResult, internal::InternalError};
+pub use identity::{wire::WireIdentity, x509::X509Identity};
 pub use metadata::{BuildMetadata, build_metadata, version};
